remapeo J.Carrenca

+++ src/java/cl/cesfam/ENTITY/DetallePartida.java
@@ -1,7 +1,8 @@
 package cl.cesfam.ENTITY;
-// Generated 14-04-2017 22:03:42 by Hibernate Tools 4.3.1
+// Generated 16-04-2017 15:22:40 by Hibernate Tools 4.3.1
 
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -10,22 +11,22 @@ import java.util.Date;
 public class DetallePartida  implements java.io.Serializable {
 
 
-     private long idDetallePartida;
+     private BigDecimal idDetallePartida;
      private Medicamento medicamento;
      private Partida partida;
-     private long cantidad;
+     private BigDecimal cantidad;
      private Date fechaIngreso;
 
     public DetallePartida() {
     }
 
 	
-    public DetallePartida(long idDetallePartida, long cantidad, Date fechaIngreso) {
+    public DetallePartida(BigDecimal idDetallePartida, BigDecimal cantidad, Date fechaIngreso) {
         this.idDetallePartida = idDetallePartida;
         this.cantidad = cantidad;
         this.fechaIngreso = fechaIngreso;
     }
-    public DetallePartida(long idDetallePartida, Medicamento medicamento, Partida partida, long cantidad, Date fechaIngreso) {
+    public DetallePartida(BigDecimal idDetallePartida, Medicamento medicamento, Partida partida, BigDecimal cantidad, Date fechaIngreso) {
        this.idDetallePartida = idDetallePartida;
        this.medicamento = medicamento;
        this.partida = partida;
@@ -33,11 +34,11 @@ public class DetallePartida  implements java.io.Serializable {
        this.fechaIngreso = fechaIngreso;
     }
    
-    public long getIdDetallePartida() {
+    public BigDecimal getIdDetallePartida() {
         return this.idDetallePartida;
     }
     
-    public void setIdDetallePartida(long idDetallePartida) {
+    public void setIdDetallePartida(BigDecimal idDetallePartida) {
         this.idDetallePartida = idDetallePartida;
     }
     public Medicamento getMedicamento() {
@@ -54,11 +55,11 @@ public class DetallePartida  implements java.io.Serializable {
     public void setPartida(Partida partida) {
         this.partida = partida;
     }
-    public long getCantidad() {
+    public BigDecimal getCantidad() {
         return this.cantidad;
     }
     
-    public void setCantidad(long cantidad) {
+    public void setCantidad(BigDecimal cantidad) {
         this.cantidad = cantidad;
     }
     public Date getFechaIngreso() {
